tried out different settings

+++ src/LFNS/LFNSSettings.h
@@ -67,7 +67,7 @@ namespace lfns {
         bool use_premature_cancelation = false;
         int print_interval = 1;
         int acceptance_info_print_interval = 100;
-        int rejection_quantile_for_density_estimation = 0.001;
+        double rejection_quantile_for_density_estimation = 0.001;
         int H = 200;
     };
 };
